refactor: split my routes tab section
